ai-service-vision/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.10.6
pydantic-settings==2.1.0
openai>=1.30.0
PyMuPDF==1.23.7
Pillow==10.1.0
